Handle non-readiness without throwing exceptions

+++ src/MessageHandler/TerminateMachineMessageHandler.php
@@ -4,14 +4,17 @@ declare(strict_types=1);
 
 namespace App\MessageHandler;
 
+use App\Enum\MessageHandlingReadiness;
 use App\Event\MachineTerminationRequestedEvent;
-use App\Exception\MessageHandlerNotReadyException;
 use App\Exception\RemoteJobActionException;
 use App\Message\TerminateMachineMessage;
 use App\ReadinessAssessor\ReadinessAssessorInterface;
 use Psr\EventDispatcher\EventDispatcherInterface;
+use Psr\Log\LoggerInterface;
 use SmartAssert\WorkerManagerClient\Client as WorkerManagerClient;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
+use Symfony\Component\Messenger\Exception\ExceptionInterface;
+use Symfony\Component\Messenger\MessageBusInterface;
 
 #[AsMessageHandler]
 final readonly class TerminateMachineMessageHandler extends AbstractMessageHandler
@@ -20,17 +23,22 @@ final readonly class TerminateMachineMessageHandler extends AbstractMessageHandl
         private WorkerManagerClient $workerManagerClient,
         EventDispatcherInterface $eventDispatcher,
         ReadinessAssessorInterface $readinessAssessor,
+        MessageBusInterface $messageBus,
+        LoggerInterface $logger,
     ) {
-        parent::__construct($eventDispatcher, $readinessAssessor);
+        parent::__construct($eventDispatcher, $readinessAssessor, $messageBus, $logger);
     }
 
     /**
      * @throws RemoteJobActionException
-     * @throws MessageHandlerNotReadyException
+     * @throws ExceptionInterface
      */
     public function __invoke(TerminateMachineMessage $message): void
     {
-        $this->assessReadiness($message);
+        $readiness = $this->assessReadiness($message);
+        if (MessageHandlingReadiness::NOW !== $readiness) {
+            return;
+        }
 
         try {
             $machine = $this->workerManagerClient->deleteMachine($message->authenticationToken, $message->getJobId());
